refactor(iroh-net): remove unused `expired` field from `Endpoint` (#1484)
## Description

This field is unused and models a concept from tailscale that we don't
have. This doesn't mean the peer's info is stale, or that it was last
seen long ago, it means that the peer's public key expired based on an
explicit `KeyExpiry *time.Time`.

From their docs

>```
>// Expired is whether this node's key has expired. Control may send
>// this; clients are only allowed to set this from false to true. On
>// the client, this is calculated client-side based on a timestamp sent
>// from control, to avoid clock skew issues.
>```

In tailscale the `Expired` is only set by a `ExpiryManager` based on
that timestamp, from local computation; and only when the `NetworkMap`
has not updated it based on a `MapResponse` requested to the server. We
have none of this concepts.

If we want to define that a peer/endpoint expired, we should do it on
our terms later on if the need arises.

## Notes & open questions

none

## Change checklist

- [x] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.