feat(sync): Add `close_replica` and enable subscriptions without active sync (#1495)
## Description

Based on #1491 

This adds two related features that were missing so far:

* Add `iroh_sync::store::close_replica` to remove a replica instance
from the store-internal cache of open replicas. This is needed because
otherwise it grows unbounded (for anchor nodes especially).
* Properly handle open replicas in the `LiveSync`, and through this
allow to subscribe to replica events for replicas that are not in the
set of actively syncing replicas. This was not possible before (you
could only subscribe to replicas that are active for sync)
* Add test for subscribing without sync

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.