Console download progress (#1526)
## Description

This brings the console download GUI in line with iroh get. You get a
nice progress bar instead of just an event dump.

In addition I fixed(?) a few other things. The different download stages
1/3 connecting, 2/3 requesting 3/3 downloading are now only shown
transiently as the progress bar message. The only thing that is
permanent is "requesting <hash>" and the (successful or unsuccessful)
outcome.

To make this possible I had to change the test regexes to only look for
the permanent output, which I think is not bad. We really don't want to
test the transient outputs in all the tests. Maybe there should be one
test that does that that is specifically made to test the "console UI".

This is one thing I found that is related to
https://github.com/n0-computer/iroh/issues/1498

## Notes & open questions

For debugging the event dump is not that bad. Maybe we need a mode where
you always get the event dump?

## Change checklist

- [x] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.