Gather statistics for transfer_collection (#1546)
## Description

Add some very detailed stats about where exactly the transfer_collection
fn spends it's time.
DRY the progress reporting in get, to make sure the same code is used
when piping to console and when writing to a file

## Notes & open questions

We got lots of info now, but where exactly should it go. Events?

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.