Remove custom get and custom collections (#1575)
## Description

As discussed, this removes the custom get request. Fixes
https://github.com/n0-computer/iroh/issues/1554

A custom get request is going to be replaced by the following:

- create a connection under another ALPN (not the iroh-bytes one)
- do whatever custom exchange needed to figure out what is to be done
- use the existing quinn stream for a bao sync

Also removes the custom collection parsing. Collections are now just
sequences of blake3 hashes. Fixes
https://github.com/n0-computer/iroh/issues/1553

Custom collections are going to be replaced by two things: For the
purpose of the network protocol, they are going to be replaced with an
exchange under a custom ALPN as described above. For the purpose of GC
they are going to be replaced by a way to enumerate live nodes.

## Notes & open questions

How detailed will the examples have to be? It is hard to think of a
small example that does not seem arbitrary, but a big example would be a
lot of work.

Here is a possible big example: We define a protocol that allows
exchanging arbitrary cids, with arbitrary hash fn. We could call it
swapbits. We could even implement a more complex protocol to exchange
dags, we could call it syncgraph.

Then we implement a node that supports this custom protocol and does the
actual exchange using bao.

But as you can imagine this would be a rather large example.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.