fix(`iroh::downloader`): remove hash from `providers` in two missed cases (#1584)
## Description

We weren't removing the hash of a request that was scheduled, became
ready but couldn't be executed. This caused getting the next hash for a
provider to pop a hash no longer necessary.

I added an invariant check for the provider map and it seems to be we
missed a couple removals in #1480 so this fixes removal in cancellation
as well

## Notes & open questions
one small change, seemingly unrelated, is that the `debug_assert` now
uses the debug version of the hash instead of display. This is because
throughout the file we log the kind in debug, which in turn uses the
debug impl for the hash. This difference made it a bit harder to follow
the logs because the printed hash didn't match those in the rest of the
logs. The change makes this consistent

## Change checklist

- [x] Self-review.
- [ ] Documentation updates if relevant.
- [x] Tests if relevant.