fix(derper): update config to auto generate keys (#1599)
## Description

Deploying a derper node is a bit cumbersome right now if you don't run
with default config. You have to provide a `secret_key` but have no easy
path to generate one so either you dance to create a fresh config and
then modify it or need to use some external thing to generate keys for
you. Either way it's cumbersome.

This just adds an extra handler to inject a secret key if not present
already in the supplied config.

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.