Enable GC (#1604)
## Description

Enable at an interval of 5 minutes plus the actual GC time.

## Notes & open questions

If somebody has data in their iroh data dir from 2 versions ago, this
will delete all the data since it is not tagged. Is that acceptable?

Data added 1 version ago would have tags, so it would be fine. Also, of
course data that has been added by reference will not be touched.

So I think it is OK.

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.