Fix flakiness of gc_hashseq test (#1609)
## Description

The issue was that a temp tag was not inserted into the live set. This
is not a problem if it happens outside a gc. But if it happens right
between a gc mark phase and a gc sweep phase, the temp pin will not be
considered, and the data associated with the temp pin might be deleted.

I knew that this had to be done, and I think I even did it at some
point, but it must have been lost in some refactoring.

## Notes & open questions

There are a number of places where temp pins are not used. See
https://github.com/n0-computer/iroh/issues/1611 . But I think this
should be done in a separate PR since this definitely fixes something.

This increases the gc frequency so much that gc is running continuously
with no pause whatsoever for some tests. That was the only way to get
the failure mentioned above at least semi-reliably.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.