fix(iroh-net): reverse ip-port mapping stores only direct addresses in the peermap (#1606)
## Description

We were storing reverse mappings for `SendAddr::Derp` in the `PeerMap`.
This means basically treating the DERP as a peer itself. We concluded
this is something we shouldn't do. To fix the issue, the better solution
is to enforce it via the type system. As such, we now only do reverse
mapping from socket addresses.

As a note, I added a new type `IpPort`. Using `SocketAddr`s as key to
hashmaps often goes wrong with ipv6 because of the flow control & class
fields it contains. These fields can't be assumed stable even within the
same connection: sending a the socket address made from an (ipv6, port)
and receiving a response can make it look like the socket address is
different (same ip, same port, different flow labels) Ignoring these
fields has some limitations in extremely exceptional cases (link local
addresses with a single interface) As a note, ignoring these fields is
what tailscale does (all they store is an ip and a port). This makes
communications over ipv6 work better in the common case so I think it's
worth doing

## Notes & open questions

Some references
- flow label [RFC 2460 section
6](https://datatracker.ietf.org/doc/html/rfc2460#section-6)
- traffic class [RFC 2460 section
7](https://datatracker.ietf.org/doc/html/rfc2460#section-7)
- scope id [RFC 4007](https://www.rfc-editor.org/rfc/rfc4007)


## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [ ] Tests if relevant.