refactor(iroh-sync): remove generic from SyncEngine (#1648)
## Description

With #1612 we moved all operations on replicas into a dedicated actor,
with all communication happening over a channel. A nice consequence of
this is that the iroh-sync store only lives in the actor thread, and
therefore we don't need a generic on the `SyncEngine` and the `Node`
anymore. This PR removes the generic.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.