fix(iroh-net): temp fix for progress bar when downloading a hash seq (#1658)
## Description

Fixes the download progress bar when downloading a hash seq where the
size is not known.

## Notes & open questions

Is this good enough? The only way to get a **proper** progress for a
hash seq where you don't know the overall size is to do a verified size
request before and give the sum of the sizes to the progress logic. But
that would be another roundtrip. Fine for GUI stuff, but in many cases
we don't really care, so I would hate to do this all the time.

## Change checklist

- [x] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.