iroh-net: Use BTreeSet for AddrInfo (#1672)
## Description

rationale: this way the serialized representation of AddrInfo and all
the things that contain AddrInfo is unique instead of constantly
toggling when there are 2 or more addrs.)

## Notes & open questions

Is there any place that will only work if the addrs are randomized? I
don't think so, and I also don't think we should have it. But you never
know...

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.