feat(iroh-sync): Sync propagation (#1613)
## Description

This PR implements sync propagation. 

* When a set-reconcilation sync completes, `iroh-sync` now returns a
struct `AuthorHeads` which contains a map with an entry for each author
for which we *received* entries in the sync operation, and the *latest
timestamp* of all entries we received. So it is a kind of "cheapo vector
clock", indicating "what's the latest entry we received in this sync op
for each author"
* This `AuthorHeads` struct is then sent as a gossip message * to
neighbors only* as a `SyncReport`
* When receiving a sync report, a node checks if the `AuthorHeads` have
any news to offer (i.e. if it contains a timestamp for an author newer
than the latest existing entry for this author in the local store). If
so, the node starts a sync request *to the node we received the report
from*

Through this, the sync operations should propagate, and through checking
the `AuthorHeads` if there are actually news for us loops should be
avoided.

The PR also contains a much-needed refactoring for better state handling
in the `live` actor.

## Notes & open questions

My biggest question is if that propagation model of sending
the`AuthorHeads` to neighbors will scale in practice.

There is a test included that passes mostly, but sometimes not on CI
(timeout of 60s reached before receiving all events). It really should
not take 60s to sync 10 nodes together on localhost.

Some notes:
* For big sync operations with many messages we might want to limit the
`AuthorHeads` to some number of authors
* We will want to debounce and merge sending sync reports over some
interval
* We might want to delay sending sync reports, or not send sync reports
for very recent entries,to not "race" with gossip


<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.