refactor(iroh-net): split `endpoint` module and reduce `PeerMap` surface (#1718)
## Description

Refactors to prepare for *actual* refactors of `PeerMap`:

* Split the `endpoint` module into `peer_map` and `endpoint` to give
them a managaeable size
* Access to `PeerMap` is now all through methods, no more direct access
to the `PeerMapInner` and `Endpoint`s in there, `PeerMap::read` and
`PeerMap::write` are removed
* Move some pure logic from `MagicSockInner` to `PeerMap`

This should contain *no* functional changes, only pure refactoring. IMO
this gives us a better structure to start with actually refactoring
`PeerMap` and `Endpoint` to be less bug-prone (through properly
abstracted data structures with a limited API surface).

Best way to review is to go commit-by-commit.


## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.