refactor(iroh-net): improve `PeerMap` and `Endpoint` abstractions (#1724)
## Description

* refactor: Make the abstractions around `PeerMap` and `Endpoint` more
solid:
* Better getter API on `PeerMap`, making sure that no mismatch between
the indexes and endpoints can emerge
* No more public acccess to endpoint fields from the peermap (only
methods)
* fix: Improvments to tracing spans, fields, levels
* fix: Actually send out pings after handling call-me-maybe messages (we
did not do that)
* tests: add a test that recreates a node in a loop and connects to the
same server node over derp

## Notes & open questions

The new test passes for me locally but still is super slow (~40 seconds
for a loop of 5 times bind & connect & transfer 20 bytes). Something is
still not correct. It should never take 10 seconds to connect and
transfer 20 bytes.

I think I need some help here - my guess is now that something in the
derper is not correct, but not sure.
<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.

---------

Co-authored-by: dignifiedquire <me@dignifiedquire.com>