fix: do not wait_idle on endpoint close (#1753)
## Description

In a longer debugging session with @dignifiedquire we found out that
`wait_idle` will *always* wait on a close timer for a few 100 ms before
actually completing. If application layer protocols always take care of
finishing their work (which we do) this is not needed at all. Also the
magicsock will very likely stay around long enough to still send out the
close frame.

This together with #1752 brings the runtime of the `derp_connect_loop`
test down from 15s to 4s on my machine! First round is 2-3s (due to bugs
in netcheck making it too slow likely) and subsequent round are all
>500ms!

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.