fix(iroh-sync): prevent panic in namespace migration (#1775)
## Description

The migration as commited with #1770 panics when starting iroh with an
existing data dir. The reason is that redb panics when deleting table in
a transaction where they were opened as well. See:
https://github.com/cberner/redb/issues/715

This fixes this by moving the table deletion to a separate transaction.

The limitation and potential panic in redb was fixed in
https://github.com/cberner/redb/pull/716 so once we upgrade to the next
(still unreleased) version of redb, this can be removed again.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.