feat: Add ability to connect just by node id
This adds a fn to the magicsock to connect just by node id. Since a node
id is insufficient to connect, a configurable discovery mechanism is
added that allows finding direct addresses or at least the region id for
a node id.

An example exercising the discovery mechanism has been added in
https://github.com/n0-computer/iroh-examples/pull/2

- [x] Self-review.
- [x] Documentation updates if relevant.
- [ ] Tests if relevant.