Expose endpoint (#1790)
## Description

Expose the endpoint on the iroh node for some advanced use cases, as
discussed yesterday.

## Questions and remarks

Question: There are some fns on the node that just forward fields from
the endpoint. Should they be removed now that the endpoint is public, or
removed and moved into the NodeClient?

```rust

    /// Return the [`NodeAddr`] for this node.
    pub async fn my_addr(&self) -> Result<NodeAddr> {
        self.inner.endpoint.my_addr().await
    }

    /// Get the DERP region we are connected to.
    pub fn my_derp(&self) -> Option<u16> {
        self.inner.endpoint.my_derp()
    }
```

Remark: Also removed a type parameter bound that was not needed and
looked weird.

---------

Co-authored-by: Friedel Ziegelmayer <me@dignifiedquire.com>