refactor(iroh): add `--start` option to CLI commands (#1802)
## Description

Changes to CLI commands:

* All RPC commands now have a `--start` option that starts a node, runs
the command, and then shuts down the node
* `iroh get` is now `iroh blob get`, and needs `--start` flag to run
standalone
* `iroh start SOURCE` is now `iroh start --add SOURCE`
* Improvements to the node startup code and error handling

## Notes and open questions

We could remove the `--add SOURCE` option from `iroh start` if there was
a way to to keep running the node after starting it with a command.
Could also be something like an `--exec` option. But let's leave that to
a followup.

---------

Co-authored-by: Franz Heinzmann (Frando) <frando@unbiskant.org>
Co-authored-by: Asmir Avdicevic <asmir.avdicevic64@gmail.com>