refactor(iroh): remove the addr arg from start (#1830)
## Description

Removes the add from the start args as discussed internally. Reason for
this is that it's currently not used. In reality, the port was not used.
If we do add this back, most likely we want to also use the address to
allow ipv6 listening nodes. In the spirit of removing code we opt for
letting the code reflect how it's actually used in the moment

## Notes & open questions

na

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.

---------

Co-authored-by: Asmir Avdicevic <asmir.avdicevic64@gmail.com>