chore(ci): run cargo build for android platforms
Due to a bug in redb we are currently not running tets on android.
This is a bit problematic as we break android builds this way.

This makes sure we can at least build, so the linker will also run.

The next step would be to run the tests for android which do not
depend on redb.