chore(ci): run cargo build for android platforms (#1843)
## Description

Due to a bug in redb we are currently not running tets on android.
This is a bit problematic as we break android builds this way.

This makes sure we can at least build, so the linker will also run.

## Notes & open questions

The next step would be to run the tests for android which do not
depend on redb.
 
Running `cargo check` in a separate step does not provide any benefit,
it is an extra compiler step which does take time and does not result
in catching errors which otherwise aren't caught.  Maybe we should also
remove this step from the non-cross checks, but that belongs better in
a different PR probably.

## Change checklist

- [x] Self-review.
- ~~Documentation updates if relevant.~~
- ~~Tests if relevant.~~