feat(bytes): switch to a single directory for the flat store (#1855)
## Description

Have the constructor for the bao db take a single directory. Internally
it will still have 3 directories, although this is not strictly speaking
necessary because the files can be distinguished by their extension.

Implements https://github.com/n0-computer/iroh/issues/1662

## Notes & open questions

~This does not yet include a migration. Do we need one, or do we declare
that the number of changes is just too big for this release? It's not
super hard to write, but we would then have to drag it around for a
while.~

Added a migration

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.