chore(ci): Configure sccache to use local caches
Because we use self-hosted runners we do not want the sccache action
to query the github action cache service for each compilation request.
Instead it should be using the local cache.

This means the first time something is compiled on a runner it will
not yet be in the cache, even if another runner already did compile
this.  However because the cache size is set reasonably large over
time all the common dependencies will end up cached on each
self-hosted runner.

Maybe it will be possible to run a self-hosted caching service
sometime, but for now relying on a large local cache is expected to be
reasonable.