refactor(logging): Reduce loglevel of blob GC (#1866)
## Description

The blob GC was logging at info level, filling the info-level log with
tons of repeated messages that are pretty useless.  They probably all
belong on debug, or maybe even below.

## Notes & open questions

Possibly some of these might be info if they carry something useful,
like showing a number that's not 0 or so.  But as a first
approximation reducing everything to debug seems about right.

## Change checklist

- [x] Self-review.
- ~~Documentation updates if relevant.~~
- ~~Tests if relevant.~~