refactor: client `Entry` with methods to read content (#1854)
## Description

This adds a `iroh::client::Entry` struct which wraps the
`iroh_sync::Entry` and adds methods to read the content of the entry.
Reading the entry needs a reference to either a `Doc<C>` or the
`Iroh<C>` client (both work equally).

@flub mentioned that this is missing from the expected API. I started
this at the hack session but didn't get very far but now it's in a good
shape IMO.

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.

---------

Co-authored-by: Friedel Ziegelmayer <me@dignifiedquire.com>