chore(ci): Use TRACE logging for tests run by nextest
The default log-level being set to INFO makes it really difficult
to debug flaky tests on CI.  This is a regression from earlier where
we had TRACE-level logging for tests using
iroh_test::logging::setup(), which was only possible for
single-threaded executor tests.

Now that we are running with cargo-nextest per-test output capturing
also works seamlessly for multi-threaded runners because of the
cargo-nextest execution model.  This means we can set a detailed log
level again by default, thus giving us detailed logs on any failure
without having to reproduce it.