chore(ci): Use TRACE logging for tests run by nextest (#1902)
## Description

The default log-level being set to INFO makes it really difficult
to debug flaky tests on CI.  This is a regression from earlier where
we had TRACE-level logging for tests using
iroh_test::logging::setup(), which was only possible for
single-threaded executor tests.

Now that we are running with cargo-nextest per-test output capturing
also works seamlessly for multi-threaded runners because of the
cargo-nextest execution model.  This means we can set a detailed log
level again by default, thus giving us detailed logs on any failure
without having to reproduce it.

## Notes & open questions

As can be seen in the doctests changes, where we can not use nextest,
we could also do this in some other ways.  Given we now depend on
cargo-nextest I guess we can as well go for always setting TRACE level
as the CI configuration stays much simpler.

We could also take this much further and simplify the machinenry of
iroh-test::logging to not jump through all those hoops.  Though I guess
people might still run the test suite with `cargo-test` which would
be nicer if we leave it in.

## Change checklist

- [x] Self-review.
- ~~Documentation updates if relevant.~~
- ~~Tests if relevant.~~