Key hex parse (#1904)
## Description

Especially for tests I find hex keys nicer. So this does to PublicKey
and PrivateKey what we are already doing with Hash: allow parsing as
either base32 or hex. There is no potential for mixing these up, since
the length will be different except for the case of 1 or 2 bytes, and we
got 32 bytes here.

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.