Fix partial downloads for individual blobs (#1903)
## Description

To check if data is available, it is not enough to call get_partial. We
now have get_possibly_partial which returns all 3 possible cases
complete/partial/missing.

contains was renamed to entry_status and removed to PartialMap.
get_partial was replaced with get_possibly_partial and extended as
described above.

Unfortunately a bit of repetition due to the code duplication in the
downloader.

## Notes & open questions

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.