fix(iroh-net): Try fix flaky udp_blocked test
This test is flaky because somehow the executor is entirely blocked
for a solid 5 seconds after which the entire netcheck report is timing
out.  Thus the part of the netcheck report that is being tested never
executes.

What blocks the executor however is a mystery to me.  This fix assumes
that it is generating the backtrace in the log, because there simply
isn't anything else that it can be.  Thus the relevant change is no
longer producing the backtrace.  The error message is already pretty
obvious.

If this isn't enough to fix this then we can try spawning the captive
portal check as a task.  But we'd still be blocking the executor,
which just isn't any good.

This also tweaks the error reporting to not generate so much logging
noise when the captive portal fails in an expected way.