Add an additional event FoundLocal. (#1911)
## Description

This event needs to be considered when computing how much data needs to
be downloaded in case of resume.

Also change hashseq download to assume that at least the hashseq itself
is complete. We currently do not support resume if the download of the
hashseq itself was interrupted.

## Notes & open questions

1. I think this provides the required info to get good progress bars
even in case of resume. But I don't know for sure until I have hacked it
into sendme.
2. Sendme does 2 roundtrips and has accurate progress bars. Iroh does 1
request and has a 2 level progress bar because the total size is not
known in advance. Should we change iroh to do what sendme does?

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [ ] Tests if relevant.