feat: additional public get utils
This adds some more higher level utilities for getting data via
iroh-bytes.

get::db adds utilities to get into a store, including considering data
that is already present locally
get::request adds utilities to do non-trivial iroh-bytes requests such
as asking for the verified and unverified size.

## Notes & open questions

Note: some of this is not yet completely finished, but I think it is
good to have it nonetheless. It is used in sendme and in the content
tracker example as well as the iroh gateway.