chore(ci): Do not test nightly rust on every PR (#1940)
## Description

Normal CI jobs for PRs now only run using the stable rust compiler,
this gives fewer surprises to users.

Instead the beta compiler is run in a new workflow which runs once a
week.  When the beta workflow fails we get a post in a discord channel.

## Notes & open questions

The required checks for branch protection have changed again.  We no
longer need the nightly checks in those.

## Change checklist

- [x] Self-review.