WIP: Send a call-me-mabye from the acceptor sooner
There is a bug that an accepting node does not send a
call-me-maybe to the connecting node for a long-long time.  Then the
connecting node never knows which direct addresses to try.

The accepting node should send call-me-maybe as soon as it is trying
to send something.  However the exact condition in this WIP code might
not be correct yet.