refactor(iroh-net): Introduce a minimal DerpUrl
This addresses two issues for now: some very minimal normalisation and
nicer Debug printing.

The type is purely a wrapper and does extremely minimal extra things
to the underlying Url.  There is no validation happening and no added
contraints to the URLs.  It is still possible to construct a DerpUrl
that will fail, e.g. by using the mailto: scheme.  The users of the
DerpUrl are expected to fail in such cases.