refactor(iroh-net): Introduce a minimal DerpUrl (#1993)
## Description

This addresses two issues for now: some very minimal normalisation and
nicer Debug printing.

The type is purely a wrapper and does extremely minimal extra things
to the underlying Url.  There is no validation happening and no added
contraints to the URLs.  It is still possible to construct a DerpUrl
that will fail, e.g. by using the mailto: scheme.  The users of the
DerpUrl are expected to fail in such cases.

## Notes & open questions

I was initially trying to keep this crate-local to iroh-net.  However
iroh-net has a surprisingly large public API and that would be painful
and largely defeat the point to leave all those public APIs as Url.
Thus it is a public API.

As a consequence there are a few uses of this outside of the iroh-net
crate as well.  I think they are reasonable.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.