clean up should ping logic
The decision on whether to send pings was taken in two places which
were both replicating the same logic and had to agree in the right
way.  This logic could be taken entirely in .want_full_ping() which
leaves the address selection logic just concerned about address
selection.  Much simpler to reason about without duplicate logic.