refactor: download and export structure and progress events (#2003)
## Description

This PR has a number of cleanups regarding download and export
functionality:
* We currently have an enum `DownloadProgress` which contains events for
both downloading a blob and for exporting a blob from the node's blob
store to files on the local filesystem. This PR moves the export
progress events into an `ExportProgress` enum.
* The code to export blobs to the local file system is moved from `iroh`
to `iroh-bytes`
* Code duplication around export progress is reduced and a couple of RPC
and client structs are cleaned up accordingly

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [ ] Tests if relevant.