refactor: Fallible store traits (#2005)
## Description

This makes the store traits fallible, which is a precondition for wiring
up a store that uses IO for the blob index, such as the redb store.

There will be more changes to the traits, but this is necessary in any
case and also a bit painful, so let's get it out of the way.

This also needs to extend the list-x responses with an error case,
unfortunately.

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [ ] Documentation updates if relevant.
- [x] Tests if relevant.