Slow down sending of call-me-maybe
We now only send these on stayin-alive heartbeats.  That is every 5
seconds.  Due to how the setup works this means worst case the
acceptor stays on relay for 2 more heartbeats (10+ seconds) because
on the first stayin-alive it recently sent pings, but they got stopped by
a NAT firewall.  Only the 2nd stayin-alive results in actual pings
going through.

So this brings us back to:

- no flooding call-me-maybes
- slow direct connection establishment on the acceptor

Next we'll speed this up by sending a ping on receipt of a ping.